+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- New `/contests` command!
 ### Fixed
 - Discombobulated the combobulator.
 
@@ -284,7 +286,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 [Unreleased]: https://github.com/jaytotheay/qrm3/compare/v3.1.1...HEAD
 [3.1.1]: https://github.com/jaytotheay/qrm3/releases/tag/v3.1.1
-[3.1.0]: https://github.com/jaytotheay/qrm3/releases/tag/v3.1.0
 [3.0.0]: https://github.com/jaytotheay/qrm2/releases/tag/v3.0.0
 [2.9.2]: https://github.com/miaowware/qrm2/releases/tag/v2.9.2
 [2.9.1]: https://github.com/miaowware/qrm2/releases/tag/v2.9.1